+++ CHANGELOG.md
@@ -26,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Inline Function was breaking code in some scenarios, which has been fixed. Props to @qpwo for raising it up 🙏
 
-## [5.4.2]
+## [5.4.2] - 2021-07-06
 
 ### Fixed